refactor media components
