+++ enjoy/src/renderer/components/audios/audios-segment.tsx
@@ -4,7 +4,7 @@ import {
   AppSettingsProviderContext,
 } from "@renderer/context";
 import { Button, ScrollArea, ScrollBar } from "@renderer/components/ui";
-import { AudioCard, AddMediaButton } from "@renderer/components";
+import { AudioCard, MediaAddButton } from "@renderer/components";
 import { t } from "i18next";
 import { Link } from "react-router-dom";
 
@@ -62,7 +62,7 @@ export const AudiosSegment = (props: { limit?: number }) => {
 
       {audios.length === 0 ? (
         <div className="flex items-center justify-center h-48 border border-dashed rounded-lg">
-          <AddMediaButton />
+          <MediaAddButton />
         </div>
       ) : (
         <ScrollArea>
