+++ enjoy/src/renderer/components/medias/index.ts
@@ -1,15 +1,6 @@
 export * from "./media-left-panel";
 export * from "./media-right-panel";
 export * from "./media-bottom-panel";
-export * from "./media-player-controls";
-export * from "./media-info-panel";
-export * from "./media-recordings";
-export * from "./media-current-recording";
-export * from "./media-transcription";
-export * from "./media-transcription-read-button";
-export * from "./media-transcription-generate-button";
-export * from "./media-player";
-export * from "./media-provider";
+
 export * from "./media-loading-modal";
-export * from "./add-media-button";
-export * from "./media-transcription-print";
+export * from "./media-add-button";
