+++ enjoy/src/renderer/components/medias/media-bottom-panel/index.ts
@@ -1,0 +1,4 @@
+export * from "./media-bottom-panel";
+export * from "./media-current-recording";
+export * from "./media-waveform";
+export * from "./media-player-controls";
