+++ enjoy/src/renderer/components/medias/media-bottom-panel/media-current-recording.tsx
