+++ enjoy/src/renderer/components/medias/media-bottom-panel/media-player-controls.tsx
