--- enjoy/src/renderer/components/medias/media-caption-panel/index.ts
@@ -1,4 +1,0 @@
-export * from "./media-caption-panel-tabs";
-export * from "./media-tab-content-analysis";
-export * from "./media-tab-content-note";
-export * from "./media-tab-content-translation";
