+++ enjoy/src/renderer/components/medias/media-left-panel/media-transcription-read-button.tsx
