+++ enjoy/src/renderer/components/medias/media-left-panel/transcription.template.html
