+++ enjoy/src/renderer/components/medias/media-shadow-player.tsx
@@ -1,5 +1,4 @@
-import { useEffect, useContext, useState } from "react";
-import { MediaShadowProviderContext } from "@renderer/context";
+import { useState } from "react";
 import {
   MediaLoadingModal,
   MediaRightPanel,
@@ -11,49 +10,15 @@ import {
   ResizablePanel,
   ResizablePanelGroup,
 } from "@renderer/components/ui";
-import { useAudio } from "@renderer/hooks";
 import { useDebounce } from "@uidotdev/usehooks";
 
-export const AudioPlayer = (props: {
-  id?: string;
-  md5?: string;
-  segmentIndex?: number;
-}) => {
-  const { id, md5, segmentIndex } = props;
-  const { media, setMedia, setCurrentSegmentIndex, getCachedSegmentIndex } =
-    useContext(MediaShadowProviderContext);
-
-  const { audio } = useAudio({ id, md5 });
+export const MediaShadowPlayer = () => {
   const [layout, setLayout] = useState<number[]>();
   const debouncedLayout = useDebounce(layout, 100);
 
-  const updateCurrentSegmentIndex = async () => {
-    let index = segmentIndex || (await getCachedSegmentIndex());
-    setCurrentSegmentIndex(index);
-  };
-
-  useEffect(() => {
-    setMedia(audio);
-  }, [audio]);
-
-  useEffect(() => {
-    if (!media) return;
-
-    updateCurrentSegmentIndex();
-    return () => {
-      setCurrentSegmentIndex(0);
-    };
-  }, [media?.id]);
-
-  if (!audio) return null;
-
   return (
     <>
-      <ResizablePanelGroup
-        direction="vertical"
-        data-testid="audio-player"
-        onLayout={setLayout}
-      >
+      <ResizablePanelGroup direction="vertical" onLayout={setLayout}>
         <ResizablePanel defaultSize={60} minSize={50}>
           <ResizablePanelGroup direction="horizontal">
             <ResizablePanel defaultSize={40} minSize={20}>
